Move the `TopAppBarScaffold` composable to the `ext` package and add a modifier parameter
