+++ compose-multiplatform-material/src/androidxCommonMain/kotlin/com/huanshankeji/compose/material/ext/TopAppBarScaffold.androidxCommon.kt
@@ -1,10 +1,13 @@
-package com.huanshankeji.compose.material
+package com.huanshankeji.compose.material.ext
 
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.material.Scaffold
 import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
+import com.huanshankeji.compose.material.Icon
+import com.huanshankeji.compose.material.IconButton
 import com.huanshankeji.compose.material.icons.Icon
+import com.huanshankeji.compose.ui.Modifier
 
 actual class NavigationIconScope private constructor() {
     @Composable
@@ -34,13 +37,15 @@ actual class TopAppBarActionsScope(val rowScope: RowScope) {
 @Composable
 actual fun TopAppBarScaffold(
     title: @Composable () -> Unit,
+    topAppBarModifier: Modifier,
     navigationIcon: @Composable (NavigationIconScope.() -> Unit)?,
     actions: @Composable TopAppBarActionsScope.() -> Unit,
     content: @Composable () -> Unit
 ) =
     Scaffold(topBar = {
         TopAppBar(
-            title = title,
-            navigationIcon = navigationIcon?.let { { NavigationIconScope.instance.it() } },
-            actions = { TopAppBarActionsScope(this).actions() })
+            title,
+            topAppBarModifier.platformModifier,
+            navigationIcon?.let { { NavigationIconScope.instance.it() } },
+            { TopAppBarActionsScope(this).actions() })
     }) { content() }
